Fix initialization of procState counters after a failed readSummary
The code before this change assumes that `mProcStateTimeMs`
and `mProcStateScreenOffTimeMs` are either both null or both non-null.
However, if during the readSummary operation one of the two counters
is read successfully while the other fails, we end up in
an inconsistent state.

Bug: 305894439
Test: atest PowerStatsTests
Google: 2839095
Change-Id: I7928031e78d742e196750a95f6a17795b4192530
Merged-In: I7928031e78d742e196750a95f6a17795b4192530
Signed-off-by: Dmitrii <bankersenator@gmail.com>